Create main.yml
Title: Deploy CNN Video Timer to VPS via GitHub Actions

Description:
This GitHub Actions workflow is designed to automate the deployment of the CNN Video Timer project. It is triggered upon each push to the master branch, ensuring that the latest version of the code is always running on the VPS.

Checkout Code: It starts by checking out the latest code from the master branch.
Deploy to VPS: Subsequently, it establishes an SSH connection to the VPS using a predefined private key stored securely in GitHub Secrets. It navigates to the project directory on the VPS and executes a git pull command, fetching the latest changes from the repository.
This automated deployment streamlines the development process, reducing manual intervention and the possibility of errors, ensuring that the project runs the most updated code.